# swagger.yaml
openapi: 3.0.3
info:
  title: SenAI CRM API
  description: Agentic CRM Intelligence Platform
  version: 1.0.0
  contact:
    name: SenAI CRM Team
servers:
  - url: http://localhost:8000/api
    description: Local development server
paths:
...
  /ingest:
    post:
      summary: Ingest a new email
      operationId: ingestEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailIngest'
            examples:
              standard:
                summary: Standard email
                value:
                  message_id: "msg_001"
                  sender: "customer@example.com"
                  recipient: "support@company.com"
                  subject: "Billing question"
                  body: "Hi, I have a question about my invoice."
                  timestamp: "2024-01-15T10:00:00Z"
                  headers: {}
      responses:
        '200':
          description: Email processed (new or existing)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailIngestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /status/{job_id}:
    get:
      summary: Get ingestion job status
      operationId: getJobStatus
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
                  category:
                    type: string
                  urgency:
                    type: string
                  requires_human:
                    type: boolean
                  classification:
                    type: object
        '404':
          $ref: '#/components/responses/NotFound'
...
  /respond/{email_id}:
    post:
      summary: Mark email as replied
      operationId: respondToEmail
      parameters:
        - name: email_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Email marked as replied
          content:
            application/json:
              schema:
                type: object
                properties:
                  email_id:
                    type: string
                  status:
                    type: string
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /drafts/{email_id}:
    patch:
      summary: Update draft for email
      operationId: updateDraft
      parameters:
        - name: email_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                draft_body:
                  type: string
      responses:
        '200':
          description: Draft updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  email_id:
                    type: string
                  draft_updated:
                    type: boolean
                  draft_body:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
...
  /respond/drafts/{email_id}/approve:
    post:
      summary: Approve and send a draft reply
      operationId: approveDraft
      tags: [Respond]
      parameters:
        - name: email_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Draft approved and sent
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
    ErrorEnvelope:
      type: object
      required:
        - error_code
        - message
        - details
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
    EmailIngest:
      type: object
      required:
        - message_id
        - sender
        - recipient
        - body
        - timestamp
      properties:
        message_id:
          type: string
          minLength: 1
          maxLength: 255
        sender:
          type: string
          format: email
        recipient:
          type: string
          format: email
        subject:
          type: string
          maxLength: 1000
        body:
          type: string
          maxLength: 10000
        timestamp:
          type: string
          format: date-time
        headers:
          type: object
          additionalProperties:
            type: string
    EmailIngestResponse:
      type: object
      properties:
        job_id:
          type: string
        status:
          type: string
        message:
          type: string
        existing:
          type: boolean
        email_id:
          type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error_code: VALIDATION_ERROR
            message: Missing required fields
            details: {}
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error_code: VALIDATION_ERROR
            message: Request validation failed
            details:
              body: Field required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error_code: NOT_FOUND
            message: Resource not found
            details: {}
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error_code: INTERNAL_ERROR
            message: An unexpected error occurred
            details: {}